pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sentry-test-python"
version = "1.0.0"
description = "A comprehensive test project for Sentry development and testing"
readme = "README.md"
requires-python = ">=3.8.1"
dependencies = [
    "sentry-sdk[flask]==1.40.0",
    "flask==3.0.0",
    "requests==2.31.0",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
database = [
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
]
celery = [
    "celery==5.3.4",
    "redis==5.0.1",
]
dev = [
    "pytest==7.4.3",
    "pytest-flask==1.3.0",
    "black==23.12.0",
    "flake8==6.1.0",
    "mypy==1.8.0",
]

[project.scripts]
sentry-test = "run:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest==7.4.3",
    "pytest-flask==1.3.0",
    "black==23.12.0",
    "flake8==6.1.0",
    "mypy==1.8.0",
]
